HeirloomProject-main/src/identityCapsule/backend/pyproject.toml
[tool.poetry]
name = "identity-capsule-backend"
version = "0.1.0"
description = "Identity Capsule Backend API"
authors = ["Devin <devin@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.7"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0.37"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.10.6"}
python-magic = "^0.4.27"
aiofiles = "^23.2.1"
pydantic-settings = "^2.1.0"
face-recognition = "^1.3.0"
opencv-python-headless = "^4.11.0.86"
numpy = "^2.2.2"
pillow = "^11.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
